added functionality to filter post function
